+++ src/ScottPlot5/ScottPlot5 Sandbox/Sandbox.WinForms/Form1.cs
@@ -8,11 +8,22 @@ public partial class Form1 : Form
     {
         InitializeComponent();
 
-        var mrk = formsPlot1.Plot.Add.Marker(5, 5);
-        mrk.Size = 100;
-        mrk.Color = Colors.Black.WithAlpha(0.5f);
-        mrk.MarkerLineWidth = 2;
-        mrk.MarkerLineColor = Colors.Black;
-        mrk.Shape = MarkerShape.FilledCircle;
+        double[] ys = Generate.SquareWave(5, 3);
+        double[] xs = Generate.Consecutive(ys.Length);
+
+        var sig1 = formsPlot1.Plot.Add.SignalXY(xs, ys);
+        sig1.MarkerSize = 5;
+
+        var sig2 = formsPlot1.Plot.Add.SignalXY(xs, ys);
+        sig2.Data.YOffset = 2;
+        sig2.Data.XOffset = .2;
+        sig2.ConnectStyle = ConnectStyle.StepHorizontal;
+        sig2.MarkerSize = 5;
+
+        var sig3 = formsPlot1.Plot.Add.SignalXY(xs, ys);
+        sig3.Data.YOffset = 4;
+        sig3.Data.XOffset = .4;
+        sig3.ConnectStyle = ConnectStyle.StepVertical;
+        sig3.MarkerSize = 5;
     }
 }