+++ src/ScottPlot5/ScottPlot5/Interfaces/ISignalXYSource.cs
@@ -41,7 +41,7 @@ public interface ISignalXYSource
     /// Return pixels to render to display this signal.
     /// May return one extra point on each side of the plot outside the data area.
     /// </summary>
-    Pixel[] GetPixelsToDraw(RenderPack rp, IAxes axes);
+    Pixel[] GetPixelsToDraw(RenderPack rp, IAxes axes, ConnectStyle connectStyle);
 
     /// <summary>
     /// Return the point nearest a specific location given the X/Y pixel scaling information from a previous render.